sleeptimer/pyproject.toml
[project]
name = "sleeptimer"
version = "0.1.0"
description = "Cross-platform desktop countdown timer that sleeps or locks the machine"
readme = "README.md"
authors = [
    { name = "Mridul Krishnan", email = "mridulkrishnan134@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pystray>=0.19",
    "Pillow>=10.0",
]

[project.scripts]
sleeptimer = "sleeptimer.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sleeptimer"]
